Re-organized code
