+++ setup.py
@@ -115,5 +115,5 @@ if __name__ == "__main__":
               'mod/x86_64/*',
               'mod/x86_64/.lib/*']},
           cmdclass={'build_py': build_py_mod, 'build_mod': BuildMod},
-          entry_points={'console_scripts': ['hnn-gui=hnn_core.gui:launch']}
+          entry_points={'console_scripts': ['hnn-gui=hnn_core.gui.gui:launch']}
           )
